reorganize views and update css
