+++ src/assets/layout.css
@@ -18,12 +18,18 @@
     /* background-color: beige; */
     max-width: 100%;
     width: 100%;
-    flex-grow: 1;
+    
     display: flex;
+    flex-grow: 1;
+
 
     padding: 1rem;
-    /* margin: 1rem 0 1rem 0; */
-    /* justify-content: start; */
+    margin: 2rem 0 2rem 0;
+
+}
+
+.content section{
+text-align: center;
 }
 
 section {
@@ -41,6 +47,8 @@ h3 {
     /* margin: 3rem 0 1rem 0; */
     line-height: normal;
     margin-bottom: 1rem;
+
+    text-align: center;
 }
 
 h1 {
@@ -61,8 +69,7 @@ header {
 
 footer {
     /* background-color: rgb(22, 131, 233); */
-    margin-top: 4rem;
-    margin-bottom: 2rem;
+    margin: 1rem 1rem 2rem 1rem;
 
     max-height: 100vh;
 
@@ -79,6 +86,10 @@ a:hover {
     /* background-color: hsla(0, 0%, 100%, 0.2); */
 }
 
+
+@media (hover: hover) and (pointer: fine) {
+    a:hover {     color: var(--color-accent);    }
+}
 
 
 /* links */
@@ -122,6 +133,10 @@ ol {
     max-width: 100%;
     max-height: 60vh;
     margin-bottom: 2rem;
+
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
 }
 
 .faq_category {
@@ -161,10 +176,7 @@ ol {
     gap: 2rem;
 }
 
-.details {
-    flex: 1;
-    margin-left: 1rem;
-}
+
 
 .hero_text {
     text-align: center;
@@ -180,6 +192,10 @@ li {
 }
 
 @media (min-width: 1024px) {
+    h1,h2,h3{
+        text-align: left;
+    }
+
     #app {
         max-width: 1280px;
         min-height: 100%;
@@ -219,6 +235,18 @@ li {
 
         flex-direction: column;
     }
+
+    .img_guide{
+
+    }
+    .img_guide_s{
+        display: block;
+        margin-left: 0;
+    }
+
+    .content section{
+        text-align: justify;
+        }
 
     /* Features */
     /* .item {
