+++ src/assets/menu.css
@@ -135,7 +135,7 @@ header {
       transform-origin: 0% 0%;
       transform: translate(0%, 0);
   
-      transition: transform 0.5s cubic-bezier(0.77, 0.2, 0.05, 1);
+      transition: transform 0.3s cubic-bezier(0.77, 0.2, 0.05, 1);
     }
   
   }
