--- src/views/PrivacyAppView.vue
@@ -1,144 +1,0 @@
-<script setup>
-import { Head } from '@vueuse/head'
-</script>
-
-<template>
-  <Head>
-    <title>Privacy Policy | AirSend</title>
-    <meta name="title" content="Download and install AirSend app for Android and Windows devices" />
-    <meta name="description"
-      content="Download and install the latest version of AirSend app for Android and Windows devices on Google Play Store and Microsoft Store." />
-    <meta name="keywords"
-      content="airsend download, airsend install, airsend android, airsend windows, airsend app, clipboard sync, android, windows, share clipboard to mobile, download, easy clipboard sync" />
-  </Head>
-
-  <section class="legal" id="privacy_policy">
-    <h2 id="privacy_policy">AirSend App Privacy Policy</h2>
-
-    <h3>Summary:</h3>
-    <p>The data you send via AirSend app is private and not visible to us. That data does not leave your local network.</p>
-
-    <br>
-    <h3>Long version:</h3>
-
-    <p>
-      CherryDev built the AirSend app as
-      a Free app. This SERVICE is provided by
-      CherryDev at no cost and is intended for use as
-      is.
-    </p>
-    <p>
-      This page is used to inform visitors regarding our
-      policies with the collection, use, and disclosure of Personal
-      Information if anyone decided to use our Service.
-    </p>
-    <p>
-      If you choose to use our Service, then you agree to
-      the collection and use of information in relation to this
-      policy. The Personal Information that We collect is
-      used for providing and improving the Service. We will not use or share your information with
-      anyone except as described in this Privacy Policy.
-    </p>
-    <p>
-      The terms used in this Privacy Policy have the same meanings
-      as in our Terms and Conditions, which are accessible at
-      AirSend unless otherwise defined in this Privacy Policy.
-    </p>
-
-    <br>
-    <p><strong>Information Collection and Use</strong></p>
-    <p>
-      For a better experience, while using our Service, I
-      may require you to provide us with certain personally
-      identifiable information. The information that
-      We request will be retained on your device and is not collected by us in any way.
-    </p>
-    <div>
-      <p>
-        The app does use third-party services that may collect
-        information used to identify you.
-      </p>
-      <p>
-        Link to the privacy policy of third-party service providers used
-        by the app
-      </p>
-      <ul>
-        <li><a href="https://www.google.com/policies/privacy/" target="_blank" rel="noopener noreferrer">Google Play Services</a></li>
-        <!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!----><!---->
-      </ul>
-    </div>
-
-    <br>
-    <p><strong>Log Data</strong></p>
-    <p>
-      We want to inform you that whenever you
-      use our Service, in a case of an error in the app
-      We collect data and information (through third-party
-      products) on your phone called Log Data. This Log Data may
-      include information such as your device Internet Protocol
-      (“IP”) address, device name, operating system version, the
-      configuration of the app when utilizing our Service,
-      the time and date of your use of the Service, and other
-      statistics.
-    </p>
-
-    <br>
-    <p><strong>Security</strong></p>
-    <p>
-      We value your trust in providing us your
-      Personal Information, thus we are striving to use commercially
-      acceptable means of protecting it. But remember that no method
-      of transmission over the internet, or method of electronic
-      storage is 100% secure and reliable, and We cannot
-      guarantee its absolute security.
-    </p>
-
-    <br>
-    <p><strong>Links to Other Sites</strong></p>
-    <p>
-      This Service may contain links to other sites. If you click on
-      a third-party link, you will be directed to that site. Note
-      that these external sites are not operated by us.
-      Therefore, We strongly advise you to review the
-      Privacy Policy of these websites. We have
-      no control over and assume no responsibility for the content,
-      privacy policies, or practices of any third-party sites or
-      services.
-    </p>
-
-    <br>
-    <p><strong>Children’s Privacy</strong></p>
-    <div>
-      <p>
-        These Services do not address anyone under the age of 13.
-        We do not knowingly collect personally
-        identifiable information from children under 13 years of age. In the case
-        We discover that a child under 13 has provided
-        us with personal information, We immediately
-        delete this from our servers. If you are a parent or guardian
-        and you are aware that your child has provided us with
-        personal information, please contact us so that
-        We will be able to do the necessary actions.
-      </p>
-    </div> <!---->
-
-    <br>
-    <p><strong>Changes to This Privacy Policy</strong></p>
-    <p>
-      We may update our Privacy Policy from
-      time to time. Thus, you are advised to review this page
-      periodically for any changes.
-    </p>
-
-    <br>
-    <p>This policy is effective as of 2023-03-06</p>
-
-    <br>
-    <p><strong>Contact Us</strong></p>
-    <p>
-      If you have any questions or suggestions about our
-      Privacy Policy, do not hesitate to contact us at <code>acherrydev</code> and <code>@gmail.com</code>.
-    </p>
-
-  </section>
-</template>
