+++ src/views/legal/PrivacyAppView.vue
@@ -1,0 +1,87 @@
+<script setup>
+import { Head } from '@vueuse/head'
+</script>
+
+<template>
+  <Head>
+    <title>Privacy Policy | AirSend</title>
+  </Head>
+
+  <section class="legal" id="privacy_policy">
+    <h1 id="privacy_policy">AirSend App Privacy Policy</h1>
+
+    <h2>Summary:</h2>
+    <blockquote>
+      <p>The data you send via AirSend app is private and not visible to us. That data does not leave your local network.</p>
+    </blockquote>
+
+    <br>
+    <h2>Long version:</h2>
+
+    <p>
+      If you choose to use our App, then you agree to
+      the collection and use of information in relation to this
+      policy.
+    </p>
+
+    <br>
+    <h3>Information Collection and Use</h3>
+    <p>
+      We want to inform you that whenever you
+      use our App, in a case of an error in the app
+      we collect data and information (through third-party
+      products) on your phone called Log Data. This Log Data may
+      include information such as the app version, screen size, device name, operating system version, the
+      configuration of the app when utilizing the App,
+      the time and date of your use of the App.
+    </p>
+    <br>
+    <p>Diagnostic data does not include either the information or files you transfer via the app.</p>
+
+    <br>
+    <h3>Security</h3>
+    <p>
+      No method of transmission over the Internet, or method of electronic
+      storage is 100% secure and reliable, and we cannot
+      guarantee its absolute security.
+    </p>
+
+    <br>
+
+    <h3>Changes to This Privacy Policy</h3>
+    <p>
+      We may update our Privacy Policy from
+      time to time. You are advised to review this page
+      periodically for any changes.
+    </p>
+
+    <br>
+    <p>This policy is effective as of 2023-03-06</p>
+
+    <br>
+    <h3>Contact Us</h3>
+    <p>
+      If you have any questions or suggestions about our
+      Privacy Policy, do not hesitate to contact us at <code>acherrydev</code> and <code>@gmail.com</code>.
+    </p>
+
+  </section>
+</template>
+
+<style scoped lang="css">
+blockquote {
+    margin: 0;
+}
+
+blockquote p {
+    padding: 15px;
+    background: var(--color-background-mute);
+    border-radius: 5px;
+    color:var(--color-accent)
+}
+
+p{
+  text-align: justify;
+}
+
+</style>
