+++ src/views/legal/PrivacyView.vue
